Scale of LTH logo
The LTH logo is wiser than the LU logo and shuld be scaled a bit more. This fixes the scaling of the logo. It also shifts up the title on the first page the same amount to avoid a gap. To the bigger version of the LTH logo set the option LTHlogonoscale (not included in the manual).