Support a left-foot-logo.
Dual affiliations are relatively common, so support for multiple logos
is desirable. This commit adds rudimentary support for adding another
logo to the left of the footer.

A bit more work would be desirable to make it more configurable and to
make the logic for the left/right footer more symmetric.